% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMat.R
\name{bMat}
\alias{bMat}
\title{Full interaction matrix}
\usage{
bMat(k)
}
\arguments{
\item{k}{An integer. Number of basic factors.}
}
\value{
A k by \eqn{2^{k}-1} binary matrix of all interactions between basic factors.
}
\description{
\code{bMat} returns the matrix of all the interactions between k basic factors
}
