% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{char2num}
\alias{char2num}
\title{Number representation of a generator, written as a string}
\usage{
char2num(s)
}
\arguments{
\item{s}{a non-empty string representing a generator}
}
\value{
the corresponding generator number
}
\description{
`char2num` takes a non-empty string, representing a generator, and converts it
to the equivalent generator number. In the string, each letter represents a basic factor.
}
\details{
All letters are converted to lower case to avoid confusion with upper case letters,
representing four-level factors.
The order of the letters does not matter.
Any letter is only counted once in the word.
Thus, `char2num("abcd")` and `char2num("abbcddd")` will output the same result.
}
\examples{
char2num("abd")
char2num("Addb")
}
