% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distanceDist}
\alias{distanceDist}
\title{Distance distribution of a matrix}
\usage{
distanceDist(x)
}
\arguments{
\item{x}{input matrix}
}
\value{
a vector of the distance distribution
}
\description{
\code{distanceDist(x)} computes the distribution of the hamming distances between
all pairs of rows (even identical) in matrix \code{x}.
}
\details{
For a matrix \eqn{\mathbf{X}} of size \eqn{(n \times m)}, the distance distribution is
a row-vector \eqn{B} of length \eqn{n}, where \eqn{B_i} is the number of pairs of rows
with hamming distance \eqn{i}.
Since pairs of identical rows are considered, \eqn{B_0} will always be \eqn{n}.
}
