% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hammingDist}
\alias{hammingDist}
\title{Hamming distance between two vectors}
\usage{
hammingDist(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
Hamming distance between \code{a} and \code{b}
}
\description{
\code{hammingDist(a,b)} finds the hamming distance between two vectors \code{a} and \code{b}.
}
\details{
The hamming distance is the number of elements that differs between
  two arrays.
}
