% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTable.R
\name{interaction.data}
\alias{interaction.data}
\title{Creates interaction data about all possible interactions between k factors}
\usage{
interaction.data(k)
}
\arguments{
\item{k}{number of basic factors}
}
\description{
Interaction data includes: Number, Character equivalent, Length
}
