% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchtable.R
\name{mixed_searchtable}
\alias{mixed_searchtable}
\title{Search-table for mixed-level designs}
\usage{
mixed_searchtable(m, k, p, R)
}
\arguments{
\item{m}{Number of four-level factors}

\item{k}{Number of basic factors}

\item{p}{Number of added factors}

\item{R}{Resolution}
}
\value{
A tibble containing the search-table.
}
\description{
Create a search-table for regular four-and-two-level designs.
}
\details{
Pseudo-factors are always composed of a pair of basic factors and their interaction.
}
\examples{
# Search-table for 4^1 2^6 designs in 32 runs
mixed_searchtable(1, 5, 3, 3)
}
