% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldoeR-package.R
\docType{package}
\name{mldoeR-package}
\alias{mldoeR}
\alias{mldoeR-package}
\title{mldoeR: Catalog of mixed four-and-two-level regular designs}
\description{
This packages contains all the four-and-two-level regular designs cited in the catalog.
    All designs are also characterized according to specific criteria.
}
\author{
\strong{Maintainer}: Alexandre Bohyn \email{alexandre.bohyn@kuleuven.be} (\href{https://orcid.org/0000-0001-9776-7467}{ORCID})

}
\keyword{internal}
