% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSearchTable.R
\name{mSearchTable}
\alias{mSearchTable}
\title{mSearchTable}
\usage{
mSearchTable(m, k, p, r)
}
\arguments{
\item{m}{Number of four-level factors}

\item{k}{Number of basic factors}

\item{p}{Number of added factors}

\item{r}{Resolution}
}
\value{
A tibble containing the search-table.
}
\description{
Create a search-table for mixed-level designs
}
\examples{
# Search-table for 4^1 2^6 designs in 32 runs
mSearchTable(1,5,3,3)
}
