% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2char.R
\name{num2char}
\alias{num2char}
\title{Convert a column number to a word representation}
\usage{
num2char(x)
}
\arguments{
\item{x}{Number of the interaction}
}
\value{
A character of length larger or equal to one, corresponding to a generator.
    For any number n, the length of the character will always be ceil(log2(n)).
}
\description{
\code{num2char} represents a column number as its corresponding generator.
}
\details{
Column number are always positive, non-zero integers.
Basic factors are power of two, while added factors are decomposed in powers of two and
their corresponding generator is build from it.
A generator corresponds to the basic factors used to create the levels of the added factor.
All basic factors are represented by independent lowercase letters.
}
\examples{
num2char(4)
num2char(13)
}
