% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{num2vec}
\alias{num2vec}
\title{Vector representation of a column number}
\usage{
num2vec(x, k)
}
\arguments{
\item{x}{column number to represent.}

\item{k}{(Optional) number of basic factors to consider. Default is ceil(log2(x)).}
}
\description{
`num2vec` represents the numbers presented as arguments as column vectors from
the design matrix.
}
\details{
If a numeric vector is supplied to the function, then a matrix is returned.
The cplumn size of the matrix corresponds to the size of the vector.
}
