% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchtable.R
\name{searchtable}
\alias{searchtable}
\title{Search-table for two-level designs}
\usage{
searchtable(k, p, R)
}
\arguments{
\item{k}{Number of basic factors.}

\item{p}{Number of added factors}

\item{R}{Resolution}
}
\value{
A tibble containing the search-table.
}
\description{
Search-table for two-level designs
}
