% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTable.R
\name{searchTable}
\alias{searchTable}
\title{Create the search-table for $2^{n-p}$ designs of resolution r}
\usage{
searchTable(n, p, r)
}
\arguments{
\item{n}{number of two-level factors}

\item{p}{number of added factors}

\item{r}{minimal resolution of the designs}
}
\value{
A 2^k-k-1 by p data.frame that contains the search-table
}
\description{
Create the search-table for $2^{n-p}$ designs of resolution r
}
